fix types, DTOs, refactor layer utils to core
