+++ src/core/logical_layers/types/source.ts
@@ -1,11 +1,13 @@
-interface TileSource {
+import type { LayerDetailsLegend } from '~core/logical_layers/types/legends';
+
+export interface TileSource {
   type: 'vector' | 'raster';
   tileSize: number;
   urls: string[];
   apiKey?: string;
 }
 
-interface GeoJSONSource {
+export interface GeoJSONSource {
   type: 'geojson';
   data: GeoJSON.FeatureCollection | GeoJSON.Feature;
 }
@@ -23,3 +25,33 @@ export interface LayerTileSource {
 }
 
 export type LayerSource = LayerGeoJSONSource | LayerTileSource;
+
+export interface GeoJSONSourceSourceContainer {
+  id: string;
+  source: GeoJSONSource;
+  legend: LayerDetailsLegend;
+  ownedByUser: boolean;
+}
+
+export interface TileSourceContainer {
+  id: string;
+  maxZoom: number;
+  minZoom: number;
+  source: TileSource;
+  legend?: LayerDetailsLegend;
+  ownedByUser: boolean;
+}
+export type LayerDetailsDTO = GeoJSONSourceSourceContainer | TileSourceContainer;
+
+export interface LayerSummaryDto {
+  id: string;
+  name: string;
+  source: TileSource | GeoJSONSource;
+  description?: string;
+  category?: 'base' | 'overlay';
+  group?: string;
+  copyrights?: string[];
+  boundaryRequiredForRetrieval: boolean;
+  eventIdRequiredForRetrieval?: boolean;
+  ownedByUser: boolean;
+}
