--- src/features/create_layer/utils/legendFormatter.ts
@@ -1,44 +1,0 @@
-import { convertRGBtoObj } from '~utils/bivariate/bivariateColorThemeUtils';
-import type {
-  BivariateLegendStep,
-  LayerLegend,
-} from '~core/logical_layers/types/legends';
-import type { LayerInAreaDetails } from '../types';
-
-function convertBivariateColorsToSteps(
-  colors: { id: string; color: string }[],
-): BivariateLegendStep[] {
-  return colors
-    .map(({ id, color }) => {
-      const clrObj = convertRGBtoObj(color);
-      return {
-        label: id,
-        color: `rgba(${clrObj.r},${clrObj.g},${clrObj.b},0.5)`,
-      };
-    })
-    .sort((stp1, stp2) => {
-      return stp1.label < stp2.label ? -1 : stp1.label > stp2.label ? 1 : 0;
-    });
-}
-
-export function legendFormatter(
-  details: LayerInAreaDetails,
-): LayerLegend | null {
-  if (!details.legend) {
-    return null;
-  }
-
-  if (details.legend?.type === 'bivariate') {
-    return {
-      name: details.id,
-      type: 'bivariate',
-      axis: {
-        x: details.legend.axes.x,
-        y: details.legend.axes.y,
-      },
-      steps: convertBivariateColorsToSteps(details.legend.colors),
-    };
-  }
-
-  return details.legend;
-}
