+++ src/features/layers_in_area/atoms/layersGlobalResource.ts
@@ -2,12 +2,12 @@ import { appConfig } from '~core/app_config';
 import { apiClient } from '~core/apiClientInstance';
 import { createAsyncAtom } from '~utils/atoms/createAsyncAtom';
 import { LAYERS_IN_AREA_API_ERROR } from '../constants';
-import type { LayerInArea } from '../types';
+import type { LayerSummaryDto } from '~core/logical_layers/types/source';
 
 export const layersGlobalResource = createAsyncAtom(
   null,
   async (_, abortController) => {
-    const layers = await apiClient.post<LayerInArea[]>(
+    const layers = await apiClient.post<LayerSummaryDto[]>(
       '/layers/search/global',
       { appId: appConfig.id },
       true,
